requirements.txt
joblib~=1.2.0
Flask~=2.2.2
flask-wtf~=1.0.1
python-dotenv~=0.21.0
WTForms~=3.0.1
numpy~=1.23.4
Unidecode~=1.3.6
scikit-learn==1.1.2
